requirements.txt
jinja2
rst2pdf
tabulate
pyyaml
